requirements.txt
streamlit
numpy
pillow
io
pandas
xgboost